pyproject.toml
[tool.poetry]
name = "vid-siv"
version = "0.1.0"
description = "Filters and removes unwanted or corrupt videos based on specs."
authors = ["anoduck <11767-anoduck@users.noreply.gitgud.io>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
simple-parsing = "^0.1.7"
trio = "^0.23.2"
guessit = "^3.8.0"
ffmpeg-python = "^0.2.0"
alive-progress = "^3.2.0"
python-magic = "^0.4.27"
exceptiongroup = "^1.2.2"
aioresult = "^0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
